fix: remove Prod merge to main trigger on workflow
With the auto PR behaviour creating a new VERSION file we now only want
merge to main to occur when the new VERSION file gets committed.

Previously this workflow was also being triggered when the Notify PR bot
updated the Lambda API image tag.

Also includes a fix to the auto PR workflow in how required checks are
specified.