revert: use API local cluster URLs
Switch back to using the local Kubernetes cluster URL
for API access.  This impacts which API is used by services
like admin and the document-download-api.

The reason for this revert is that a few issues have been
discovered and fixed with admin using the Lambda API and
although all appears to be working as expected, we'd like to
eliminate variables for tonight's maintenance window.