fix: combine sync secrets and merge workflows
These workflows are being combined because of an edge case that
can occur with the Production merge to main.  If there are env var
changes, the Lambda API was being updated twice in quick
succession, which was causing the second update to fail.

By combinging the workflows, the Lambda API will only be updated
once per merge event.